+++ src/components/burger-constructor/ingredient-constructor/ingredient-constructor.tsx
@@ -7,7 +7,7 @@ import styles from './styles.module.css';
 import { useDrag, useDrop } from 'react-dnd';
 import { useRef } from 'react';
 import { ItemDropTypes } from '@utils/items-drop-types';
-import { useDispatch } from 'react-redux';
+import { useDispatch } from '@utils/custom-hooks';
 import {
 	deleteIngredient,
 	swapIngredient,
