+++ src/components/feed-info/feed-to-day/feed-to-dayl.tsx
@@ -1,4 +1,4 @@
-import { useSelector } from '@services/store';
+import { useSelector } from '@utils/custom-hooks';
 import styles from './style.module.css';
 import { getTotalToDay } from '@services/ordersStream/reducer';
