+++ src/components/feed/feed-lens/feed-lens.tsx
@@ -1,5 +1,5 @@
 import { getOrders } from '@services/ordersStream/reducer';
-import { useSelector } from '@services/store';
+import { useSelector } from '@utils/custom-hooks';
 import { OrderCard } from '../../burger-orders/order-card/order-card';
 import { nanoid } from '@reduxjs/toolkit';
