+++ src/components/pages/check-list/check-list.tsx
@@ -1,7 +1,7 @@
 import { useEffect } from 'react';
 import { Container } from '../../container/container';
 import { Layout } from '../../layout/layout';
-import { useDispatch } from '@services/store';
+import { useDispatch } from '@utils/custom-hooks';
 import { connect, disconnect } from '@services/ordersStream/actions';
 import { urlWsConnect } from '@utils/url';
 import { Feed } from '../../feed/feed';
