+++ src/constants/errorsMessages.ts
@@ -1,0 +1,3 @@
+// короткий импорт его не работает, потом попробую исправить
+
+export const ORDER_ERROR_FOR_REDIRECT = 'Token is invalid';
