+++ src/utils/errorsMessages.ts
@@ -1,0 +1,1 @@
+export const ORDER_ERROR_FOR_REDIRECT = 'Token is invalid';
